requirements.txt
alpha-vantage
pandas
pytest
python-dotenv
requests